simplify settings; increase bookmark url length
