+++ docker/compose/docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.8'
-
 services:
   # PostgreSQL Database
   postgres:
@@ -44,8 +42,7 @@ services:
         condition: service_healthy
     environment:
       - DOCKER_DEPLOYMENT=compose
-      - DJANGO_SETTINGS_MODULE=pebbling.settings.docker
-      - DJANGO_ENV=dev
+      - DJANGO_SETTINGS_MODULE=pebbling.settings
       - DEBUG=True
       - SECRET_KEY=your-secret-key-here-1234567890
       - ALLOWED_HOSTS=localhost,127.0.0.1,web
@@ -53,6 +50,7 @@ services:
       - CELERY_BROKER_URL=redis://redis:6379/0
       - REDIS_URL=redis://redis:6379/1
       - DJANGO_SQLITE_MULTIPLE_DB=false
+      - LOG_LEVEL=INFO
 
   # Celery worker
   celery_worker:
@@ -68,7 +66,7 @@ services:
       - postgres
     environment:
       - DOCKER_DEPLOYMENT=compose
-      - DJANGO_SETTINGS_MODULE=pebbling.settings.docker
+      - DJANGO_SETTINGS_MODULE=pebbling.settings
       - DJANGO_ENV=dev
       - DEBUG=True
       - SECRET_KEY=your-secret-key-here-1234567890
@@ -91,7 +89,7 @@ services:
       - postgres
     environment:
       - DOCKER_DEPLOYMENT=compose
-      - DJANGO_SETTINGS_MODULE=pebbling.settings.docker
+      - DJANGO_SETTINGS_MODULE=pebbling.settings
       - DJANGO_ENV=dev
       - DEBUG=True
       - SECRET_KEY=your-secret-key-here-1234567890
