+++ pebbling/settings.py
@@ -1,0 +1,241 @@
+from pathlib import Path
+import environ
+
+# Build paths inside the project like this: BASE_DIR / 'subdir'.
+BASE_DIR = Path(__file__).resolve().parent.parent
+
+# Load environment variables
+def read_env_files(env_filenames):
+    dot_env_dir = BASE_DIR
+    for env_filename in env_filenames:
+        env_file = dot_env_dir / env_filename
+        if env_file.exists():
+            environ.Env.read_env(dot_env_dir / env_file)
+    return environ.Env()
+
+env = read_env_files([".env"])
+
+# Basic Django settings from environment variables
+LOG_LEVEL = env("LOG_LEVEL", default="INFO")
+DEBUG = env.bool("DEBUG", default=False)
+SECRET_KEY = env("SECRET_KEY", default="django-insecure-dev-key-change-in-production")
+ALLOWED_HOSTS = env.list("ALLOWED_HOSTS", default=["localhost", "127.0.0.1"])
+CSRF_TRUSTED_ORIGINS = env.list("CSRF_TRUSTED_ORIGINS", default=["http://localhost:8000", "http://127.0.0.1:8000"])
+
+# Development-specific settings
+if DEBUG:
+    INTERNAL_IPS = ["127.0.0.1"]
+
+# Production-specific settings
+if not DEBUG:
+    SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
+
+# Data directories
+DATA_BASE_DIR = BASE_DIR / env("DATA_BASE_DIR", default="data")
+SQLITE_BASE_DIR = BASE_DIR / env("SQLITE_BASE_DIR", default="data")
+
+# Application definition
+INSTALLED_APPS = [
+    "django.contrib.admin",
+    "django.contrib.auth",
+    "django.contrib.contenttypes",
+    "django.contrib.sessions",
+    "django.contrib.messages",
+    "django.contrib.staticfiles",
+    "django_celery_results",  # Stores Celery task results in DB
+    "django_celery_beat",  # Enables periodic tasks
+    "pebbling",
+    "pebbling_apps.common",
+    "pebbling_apps.users",
+    "pebbling_apps.profiles",
+    "pebbling_apps.bookmarks",
+    "pebbling_apps.feeds",
+    "pebbling_apps.home",
+    "pebbling_apps.unfurl",
+]
+
+MIDDLEWARE = [
+    "django.middleware.security.SecurityMiddleware",
+    "django.contrib.sessions.middleware.SessionMiddleware",
+    "django.middleware.common.CommonMiddleware",
+    "django.middleware.csrf.CsrfViewMiddleware",
+    "django.contrib.auth.middleware.AuthenticationMiddleware",
+    "django.contrib.messages.middleware.MessageMiddleware",
+    "django.middleware.clickjacking.XFrameOptionsMiddleware",
+]
+
+ROOT_URLCONF = "pebbling.urls"
+
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+        "DIRS": [],
+        "APP_DIRS": True,
+        "OPTIONS": {
+            "context_processors": [
+                "django.template.context_processors.debug",
+                "django.template.context_processors.request",
+                "django.contrib.auth.context_processors.auth",
+                "django.contrib.messages.context_processors.messages",
+                "pebbling_apps.common.context_processors.shift_refresh",
+                "pebbling_apps.users.context_processors.timezone_context",
+                "pebbling_apps.bookmarks.context_processors.bookmark_context",
+            ],
+        },
+    },
+]
+
+WSGI_APPLICATION = "pebbling.wsgi.application"
+
+# Logging configuration
+LOGGING = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "handlers": {
+        "console": {
+            "class": "logging.StreamHandler",
+        },
+    },
+    "root": {
+        "handlers": ["console"],
+        "level": LOG_LEVEL,
+    },
+    "loggers": {
+        "pebbling_apps": {
+            "handlers": ["console"],
+            "level": LOG_LEVEL,
+            "propagate": False,
+        },
+    },
+}
+
+# Check if we should use multiple SQLite databases
+SQLITE_MULTIPLE_DB = env.bool("DJANGO_SQLITE_MULTIPLE_DB", default=True)
+
+if SQLITE_MULTIPLE_DB:
+    _shared_sqlite_options = {
+        "ENGINE": "django.db.backends.sqlite3",
+        "OPTIONS": {
+            "transaction_mode": "IMMEDIATE",
+            "timeout": 5,
+            "init_command": """
+                PRAGMA journal_mode=WAL;
+                PRAGMA synchronous=NORMAL;
+                PRAGMA mmap_size = 134217728;
+                PRAGMA journal_size_limit = 27103364;
+                PRAGMA cache_size=2000;
+            """,
+        },
+    }
+
+    DATABASES = {
+        "default": {
+            **_shared_sqlite_options,
+            "NAME": SQLITE_BASE_DIR / "main.sqlite3",
+        },
+        "feeds_db": {
+            **_shared_sqlite_options,
+            "NAME": SQLITE_BASE_DIR / "feeds.sqlite3",
+        },
+        "celery_db": {
+            **_shared_sqlite_options,
+            "NAME": SQLITE_BASE_DIR / "celery.sqlite3",
+        },
+        "cache_db": {
+            **_shared_sqlite_options,
+            "NAME": SQLITE_BASE_DIR / "cache.sqlite3",
+        },
+    }
+else:
+    # Use single database configuration (e.g., PostgreSQL)
+    DATABASES = {
+        "default": env.db(
+            "DATABASE_URL",
+            default="sqlite:///data/main.sqlite3"
+        )
+    }
+
+# Configure database routers based on SQLITE_MULTIPLE_DB
+if SQLITE_MULTIPLE_DB:
+    DATABASE_ROUTERS = [
+        "pebbling.routers.CacheRouter",
+        "pebbling.routers.CeleryRouter",
+        "pebbling.routers.FeedsRouter",
+    ]
+else:
+    # No routers needed when using a single database
+    DATABASE_ROUTERS = []
+
+# Celery settings
+CELERY_BEAT_SCHEDULE_FILENAME = str(DATA_BASE_DIR / "celerybeat-schedule")
+
+if SQLITE_MULTIPLE_DB:
+    # Multiple database mode - use separate SQLite file for Celery broker
+    CELERY_DB_PATH = str(SQLITE_BASE_DIR / "celery.sqlite3")
+    CELERY_BROKER_URL = f"sqla+sqlite:///{CELERY_DB_PATH}"
+else:
+    # Single database mode - use Redis or other broker from environment
+    CELERY_BROKER_URL = env("CELERY_BROKER_URL", default="redis://localhost:6379/0")
+
+# Configure caches based on SQLITE_MULTIPLE_DB
+if SQLITE_MULTIPLE_DB:
+    # Multiple database mode - use database cache with separate cache_db
+    CACHES = {
+        "default": {
+            "BACKEND": "django.core.cache.backends.db.DatabaseCache",
+            "LOCATION": "django_cache_table",
+            "OPTIONS": {
+                "DATABASE": "cache_db",
+            },
+        }
+    }
+else:
+    # Single database mode - use Redis cache
+    CACHES = {
+        "default": {
+            "BACKEND": "django.core.cache.backends.redis.RedisCache",
+            "LOCATION": env("REDIS_URL", default="redis://localhost:6379/1"),
+        }
+    }
+
+CELERY_RESULT_BACKEND = "django-db"
+CELERY_ACCEPT_CONTENT = ["json"]
+CELERY_TASK_SERIALIZER = "json"
+CELERY_RESULT_EXTENDED = True
+
+# Password validation
+AUTH_PASSWORD_VALIDATORS = [
+    {
+        "NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator",
+    },
+    {
+        "NAME": "django.contrib.auth.password_validation.MinimumLengthValidator",
+    },
+    {
+        "NAME": "django.contrib.auth.password_validation.CommonPasswordValidator",
+    },
+    {
+        "NAME": "django.contrib.auth.password_validation.NumericPasswordValidator",
+    },
+]
+
+# Internationalization
+LANGUAGE_CODE = "en-us"
+TIME_ZONE = "UTC"
+USE_I18N = True
+USE_TZ = True
+APPEND_SLASH = False
+
+# Static files (CSS, JavaScript, Images)
+STATIC_URL = "static/"
+STATIC_ROOT = BASE_DIR / "static"
+STATICFILES_DIRS = [
+    BASE_DIR / "frontend/build",
+]
+
+# Default primary key field type
+DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+AUTH_USER_MODEL = "users.CustomUser"
+LOGIN_REDIRECT_URL = "profiles:index"
+LOGOUT_REDIRECT_URL = "users:login"
